// src/app/shared/types/api-types/bido-contact-dto.interface.ts
export interface BidoContactDTO {
  contactName?: string;
  firstName?: string;
  lastName?: string;
  customerName?: string;
  customerCode?: string;
  title?: string;
  PHONE1?: string;
  PHONE2?: string;
  fax?: string;
  eMail?: string;
  address?: string;
  town?: string;
  country?: string;
  primaryContact?: boolean;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}